// src/config/constants/types.ts
export interface Address {
  87?: string
  107: string
}

export interface SerializedToken {
  chainId: number
  address: string
  decimals: number
  symbol?: string
  name?: string
  projectLink?: string
}

export interface SerializedNFT {
  chainId: number
  address: string
  id: number
  baseURI: string
  symbol?: string
  name?: string
  projectLink?: string
}
// ...
interface NFTFarmConfigBaseProps {
  pid: number
  colSymbol: string
  colName: string
  colAddresses: Address
  colAddress?: string
  multiplier?: string
  isCommunity?: boolean
  image?: string
}

export interface SerializedFarmConfig extends FarmConfigBaseProps {
  token: SerializedToken
  quoteToken: SerializedToken
}

export interface DeserializedFarmConfig extends FarmConfigBaseProps {
  token: Token
  quoteToken: Token
}

export interface SerializedNFTFarmConfig extends NFTFarmConfigBaseProps {
  collection: SerializedNFT
  earningToken: SerializedToken
  stakingFee: number
  harvestingFee: number
}